# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Partner:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [Licensor, Licensee, Both]
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        contractCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePartnerRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Licensor, Licensee, Both]
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    UpdatePartnerRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Licensor, Licensee, Both]
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    ContentItem:
      type: object
      required: [id, type, title, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
        title:
          type: string
        description:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/Season'
        contractCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Season:
      type: object
      required: [id, contentItemId, seasonNumber]
      properties:
        id:
          type: string
        contentItemId:
          type: string
        seasonNumber:
          type: integer
        title:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        episodeCount:
          type: integer
          nullable: true
...
    Contract:
      type: object
      required: [id, direction, partnerId, status, endType, createdAt]
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        partnerName:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        otherTerritories:
          type: string
          nullable: true
        distributionTypes:
          type: array
          items:
            type: string
        platform:
          type: string
          nullable: true
        royaltyType:
          type: string
          enum: [revenue_share, flat_fee, other]
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        paymentTerms:
          type: string
          enum: [net_30, net_60, net_90]
          nullable: true
        notes:
          type: string
          nullable: true
        documentUrl:
          type: string
          nullable: true
        websiteLink:
          type: string
          nullable: true
        rightsInDetails:
          $ref: '#/components/schemas/RightsInDetails'
          nullable: true
        rightsOutDetails:
          $ref: '#/components/schemas/RightsOutDetails'
          nullable: true
        contentItems:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        amendments:
          type: array
          items:
            $ref: '#/components/schemas/Amendment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdByName:
          type: string
          nullable: true
    RightsInDetails:
      type: object
      properties:
        platforms:
          type: array
          items:
            type: string
        youtubeChannel:
          type: string
          nullable: true
        socialPlatforms:
          type: array
          items:
            type: string
        socialHandle:
          type: string
          nullable: true
        grantOfRights:
          type: string
          nullable: true
        exclusivityStartDate:
          type: string
          format: date
          nullable: true
        exclusivityEndDate:
          type: string
          format: date
          nullable: true
        exclusivitySameAsDuration:
          type: boolean
        marketingRights:
          type: string
          nullable: true
    RightsOutDetails:
      type: object
      properties:
        autoRenew:
          type: boolean
        hasAmendment:
          type: boolean
        exclusivity:
          type: string
          enum: [exclusive, non_exclusive]
          nullable: true
        reportingFrequency:
          type: string
          enum: [monthly, quarterly, annually]
          nullable: true
        minPaymentThreshold:
          type: number
          nullable: true
    CreateContractRequest:
      type: object
      required: [direction, partnerId, endType]
      properties:
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        otherTerritories:
          type: string
          nullable: true
        distributionTypes:
          type: array
          items:
            type: string
        platform:
          type: string
          nullable: true
        royaltyType:
          type: string
          enum: [revenue_share, flat_fee, other]
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        paymentTerms:
          type: string
          enum: [net_30, net_60, net_90]
          nullable: true
        notes:
          type: string
          nullable: true
        websiteLink:
          type: string
          nullable: true
        rightsInDetails:
          $ref: '#/components/schemas/RightsInDetails'
          nullable: true
        rightsOutDetails:
          $ref: '#/components/schemas/RightsOutDetails'
          nullable: true
        contentItemIds:
          type: array
          items:
            type: string
...
    ContractListItem:
      type: object
      required: [id, direction, status, endType, createdAt]
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        partnerName:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        distributionTypes:
          type: array
          items:
            type: string
        royaltyType:
          type: string
          nullable: true
        contentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    Amendment:
      type: object
      required: [id, contractId, date, description, createdAt]
      properties:
        id:
          type: string
        contractId:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdByName:
          type: string
          nullable: true
...
    RevenueReport:
      type: object
      required: [id, contractId, period, status, createdAt]
      properties:
        id:
          type: string
        contractId:
          type: string
        partnerName:
          type: string
          nullable: true
        period:
          type: string
        expectedDate:
          type: string
          format: date
          nullable: true
        receivedDate:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          nullable: true
        status:
          type: string
          enum: [expected, received, overdue]
        createdAt:
          type: string
          format: date-time
...
    PaginatedContracts:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContractListItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    PaginatedPartners:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Partner'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /partners:
    get:
      operationId: listPartners
      summary: List all partners
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [Licensor, Licensee, Both]
      responses:
        '200':
          description: Paginated partner list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPartners'
    post:
      operationId: createPartner
      summary: Create a new partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartnerRequest'
      responses:
        '201':
          description: Created partner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /partners/{id}:
    get:
      operationId: getPartner
      summary: Get partner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Partner details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updatePartner
      summary: Update partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartnerRequest'
      responses:
        '200':
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
    delete:
      operationId: deletePartner
      summary: Delete partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /contracts:
    get:
      operationId: listContracts
      summary: List contracts with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: direction
          in: query
          schema:
            type: string
            enum: [rights_in, rights_out]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, expired, in_perpetuity, terminated]
        - name: partnerId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: expiringWithinDays
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated contract list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContracts'
    post:
      operationId: createContract
      summary: Create a new contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Created contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reports/contracts:
    get:
      operationId: getContractReport
      summary: Get contracts report data
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [rights_in, rights_out]
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, xlsx, pdf]
            default: json
      responses:
        '200':
          description: Contract report (JSON or file)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractListItem'
                  generatedAt:
                    type: string
                    format: date-time
  /reports/expiring:
    get:
      operationId: getExpiringReport
      summary: Get expiring contracts report
      parameters:
        - name: withinDays
          in: query
          schema:
            type: integer
            default: 60
        - name: format
          in: query
          schema:
            type: string
            enum: [json, xlsx, pdf]
            default: json
      responses:
        '200':
          description: Expiring contracts report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractListItem'
                  generatedAt:
                    type: string
                    format: date-time
  /reports/royalties:
    get:
      operationId: getRoyaltyReport
      summary: Get royalty statements report
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, xlsx, pdf]
            default: json
      responses:
        '200':
          description: Royalty report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RevenueReport'
                  generatedAt:
                    type: string
                    format: date-time