# lib/api-zod/src/generated/types/paginatedPartners.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { Partner } from './partner';

export interface PaginatedPartners {
  data: Partner[];
  total: number;
  page: number;
  pageSize: number;
}